requirements.txt
Pillow
black
etcd3
ipython
pytest-pudb
requests
tenacity
